Rpi/image_processing_server/requirements.txt
absl-py==0.10.0
astroid==2.4.2
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
backports-abc==0.5
bleach==3.1.5
Brotli==1.0.7
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.5.0
colorama==0.4.3
configparser==5.0.0
cryptography==2.9.2
cycler==0.10.0
dash==1.13.4
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
docopt==0.6.2
entrypoints==0.3
Fiona @ file:///C:/Users/Yuan%20Neng/Downloads/Fiona-1.8.13-cp38-cp38-win_amd64.whl
Flask==1.1.2
Flask-Compress==1.5.0
gast==0.3.3
GDAL @ file:///C:/Users/Yuan%20Neng/Downloads/GDAL-3.1.2-cp38-cp38-win_amd64.whl
genson==1.2.1
geopandas==0.8.1
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imagezmq==1.1.1
imutils==0.5.3
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.3.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nltk==3.5
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python3==1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pdfminer3==2018.12.3.0
pickleshare==0.7.5
Pillow==8.0.0
pipreqs==0.4.10
plotly==4.8.2
plotly-express==0.4.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
pygame==1.9.6
Pygments==2.6.1
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyproj==2.6.1.post1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.1
Send2Trash==1.5.0
Shapely==1.7.0
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
sortedcontainers==2.2.2
statsmodels==0.11.1
tabpy-client==0.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9
